Jinho/func1-spring/src/main/java/paastaproj/func1spring/repositories/MemoryMemberRepository.java
package paastaproj.func1spring.repositories;

import org.springframework.stereotype.Repository;
import paastaproj.func1spring.domain.Member;

import java.util.*;

//@Repository
public class MemoryMemberRepository implements MeberRepository {

    private static Map<Long, Member> store = new HashMap<>();
    private static long sequence = 0L;

    @Override
    public Member save(Member member) {
        member.setId(++sequence);
        store.put(member.getId(), member);
        return member;
    }

    @Override
    public Optional<Member> findById(Long id) {
        //null이어도 반환
        return Optional.ofNullable(store.get(id));
    }

    @Override
    public Optional<Member> findByName(String name) {
       return store.values().stream().
                filter(member -> member.getName().equals(name)).findAny();
    }

    @Override
    public List<Member> findAll() {
        return new ArrayList<>(store.values());
    }

    public void clearStore(){
        store.clear();
    }



}
